+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/HolderPositionHandler.kt
@@ -1,10 +1,10 @@
-package com.idanatz.oneadapter.internal.animations
+package com.idanatz.oneadapter.internal
 
 import android.util.SparseIntArray
 
-internal class AnimationPositionHandler(private var highestPositionMap: SparseIntArray = SparseIntArray()) {
+internal class HolderPositionHandler(private var highestPositionMap: SparseIntArray = SparseIntArray()) {
 
-    fun shouldAnimateBind(viewType: Int, position: Int): Boolean {
+    fun isFirstBind(viewType: Int, position: Int): Boolean {
         val highestPositionForType = highestPositionMap.get(viewType, -1)
         val shouldAnimateBind = position > highestPositionForType