+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/paging/OneScrollListener.kt
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.idanatz.oneadapter.internal.utils.Logger
 import com.idanatz.oneadapter.internal.utils.extensions.findLastItemIndex
 import com.idanatz.oneadapter.internal.utils.extensions.findLastVisibleItemIndex
+import com.idanatz.oneadapter.internal.utils.extensions.isScrolling
 
 internal class OneScrollListener (
         private val layoutManager: RecyclerView.LayoutManager,
@@ -29,7 +30,7 @@ internal class OneScrollListener (
 
     // This happens many times a second during a scroll, so be wary of the code you place here.
     override fun onScrolled(view: RecyclerView, dx: Int, dy: Int) {
-        if (!isUserScrolled(view)) {
+        if (!view.isScrolling) {
             return
         }
 
@@ -64,8 +65,6 @@ internal class OneScrollListener (
         loading = false
         previousTotalItemCount = 0
     }
-
-    private fun isUserScrolled(view: RecyclerView) = view.scrollState != RecyclerView.SCROLL_STATE_IDLE
 
     private fun evaluateLoadingState(): LoadingState {
         val lastItemIndex = layoutManager.findLastItemIndex()