+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Version 2.1.2
 -------------
 * Fixed: issue #43
+* Added: isRebinding flag to onBind Metadata
 
 Version 2.1.1
 -------------